UseCase3/Subtask1/AzureFunction/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core
azure-storage-blob 
azure-eventgrid
azure-identity
azure-keyvault-secrets